ops-mesh-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-dotenv
pytest
pytest-asyncio
httpx
pytest-mock
faker
# Google Cloud Services
google-cloud-aiplatform
google-cloud-pubsub
google-cloud-storage
google-cloud-logging
google-cloud-monitoring
google-auth
google-auth-oauthlib
google-auth-httplib2
grpcio
protobuf
# Agent Protocol Dependencies
websockets
aiohttp
# A2A Protocol Dependencies
a2a-protocol==0.1.0
# Note: google-adk is implemented internally as it's not available as a public package
